requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.0
langchain-anthropic==0.1.1
langchain-community==0.0.13
chromadb==0.4.22
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
requests==2.31.0
aiofiles==23.2.1
tavily-python==0.3.0
